flight-pdsh and flight-pdsh-sync packages
`flight-pdsh` provides a copy of pdsh supporting genders, with the
default genders file at /opt/flight/etc/genders.

`flight-pdsh-sync` provides a script to create a genders file from data
held in Flight Center and copy it across the cluster.  It leverages
`flight-asset` to obtain the Flight Center data and relies on the asset
groups having a unix name set.

The java, perl and python extensions for `genders` were causing
compilation issues.  `genders` has been compiled without support for
them.  If we discover we need them, we can investigate and fix the issue
then.