Add a ruby wrapper to certificate generation
This allows for a slightly more complex CLI using options flags.
This will allow the certificate generation code to be kept separate
from the existing enable/disable https

A new hidden command has been added to the `https` script to allow
regeneration of self signed certificates. This is a stop gap measure
so the code does not need to be reorganised at this stage.

It is not part of the public interface to the command and maybe
removed without notice