Redefine the reload/restart commands for action-api
The old restart command is actually close to the concept of a
graceful "reload" as it does not stop the master puma process.
For this reason, it has been made into the "reload" command

A new "restart" command has been added which will run the "stop"
followed by "start" scripts. This is a hard reload which will
stop the master puma process