Certbot builds and installs across supported OSes
Previously, the python dependency was not embedded in the package.  We
had decided that this would probably be OK.  However, it meant that the
package would not work on Ubuntu 2004.

The python software from flight-directory is now shared between
flight-directory and flight-certbot.  This embeds the python binary in
the package (caveats apply see below).  This has resulted in python
being upgraded to 3.8 from 3.6.

There is one caveat.  Whilst flight-certbot will build on both Ubuntu
1604 and Ubuntu 2004, only the package built on Ubunt 1604 will work on
both platforms.  When built on Ubuntu 2004, the virtual env (or pipenv)
ignores the embedded python and uses the system python.  There is
probably an option to avoid this, but I have a working package for
ubuntu.