Fix race condition with an omnibus `block`
Whilst this bug may not technically be a race condition, the build
appears to fail randomly with `nginx.conf` doesn't exist

This is likely due to the `nginx` dependency not being checked out
before the `build` section runs; on software file load.

Instead a `block` section needs to be employed. This delays running
the `nginx.conf` patch until the `flight-www` software is built.
Yes this is counter-intuitive