Attempt to build howto package but back out of it
The packages containing howto guides attempt to make use of the
`Recommends: ` tag.  Unfortunately, this doesn't work for two reasons:

1. The relevant `spec.erb` files have not been updated to generate those
   tags.
2. `rpmbuild` errors if the `Recommends: ` tag is present.

It appears that whilst we can add a `Provides: flight-howto-system-1.0`
and we can add a hard dependency on it, we have no way to add a weak
dependency for RPM.

We need to abandon the idea of having weak dependencies with RPM.  They
are still likely to be useful for deb packages.