Aborted attempt at a separate RPM for file manager backend
There was an issue with building the RPM for file-manager-api after the
backend was added as a software component.  This commit adds a new
builder for the file manager backend as part of debugging that problem.

Ultimately it was not needed and will be reverted in the next commit.
The issue was due to a dependency containing `.git` submodules in its
packaged files.