Ship with a useful out-of-the-box configuration
A better (aka more secure) configuration will be needed before we ship
this, but at the moment, ease of installation, configuration and
feedback is more important.

Remove the nginx config, we don't want to allow access to the server
from outside the cluster. Having an nginx config allows that.  This will
need to be rethought before we ship.