Track package-scripts and puma scripts again
There is an issue with the `update_web_suite_package_scripts` in that it
does not pick up changes made to `postinst-configure` without `clean
--purge` being ran first.

This resulted in the services being built with incorrect postinst
scripts.  This would have been easier to see if they were tracked in
git.  So I'm reverting my decision to not have these generated files
tracked in git.

Presumably, this issue exists with `update_puma_scripts` too.