Bump build iterations for www and file-manager api
Fix issue with nginx and `404` routes.  Previously, if nginx could not
find a static file instead of responding with a `404` it responded with
a `301` redirect to the not found page.  `GET`ing the not found page
would respond with a `200` and the correct not found page content.

Now if a static page cannot be found, nginx responds with a `404` and
the content of the not found page.

End users are unlikely to be able to tell the difference.