Don't error on hunter dirs missing (info in body)
If a user logs into a Solo system before any nodes have been added to
the hunter buffer (which takes a few seconds due to Firstrun stuff) then
there will be an ls error printed for buffer and parsed dirs not
existing.

This could also be an issue in non-Solo instances where Hunter is
installed but has not been started/run/enabled yet.